requirements.txt
aspose-words
flask
joblib
matplotlib
nltk
numpy
scikit-learn==1.0.2
